# Conflicts:
#	.idea/gradle.xml
#	.idea/libraries/animated_vector_drawable_25_1_1.xml
#	.idea/libraries/appcompat_v7_25_1_1.xml
#	.idea/libraries/design_25_1_1.xml
#	.idea/libraries/espresso_core_2_2_2.xml
#	.idea/libraries/espresso_idling_resource_2_2_2.xml
#	.idea/libraries/exposed_instrumentation_api_publish_0_5.xml
#	.idea/libraries/firebase_common_9_8_0.xml
#	.idea/libraries/play_services_appindexing_9_8_0.xml
#	.idea/libraries/play_services_maps_9_8_0.xml
#	.idea/libraries/recyclerview_v7_25_1_1.xml
#	.idea/libraries/rules_0_5.xml
#	.idea/libraries/runner_0_5.xml
#	.idea/libraries/support_annotations_25_1_1.xml
#	.idea/libraries/support_compat_25_1_1.xml
#	.idea/libraries/support_core_ui_25_1_1.xml
#	.idea/libraries/support_core_utils_25_1_1.xml
#	.idea/libraries/support_fragment_25_1_1.xml
#	.idea/libraries/support_media_compat_25_1_1.xml
#	.idea/libraries/support_vector_drawable_25_1_1.xml
#	.idea/libraries/transition_25_1_1.xml
#	app/app.iml

+++ app/src/main/java/com/designproj/nickwarren/historicsj/PhotoObject.java
@@ -10,12 +10,12 @@ public class PhotoObject {
     public int idname;
     public String source;
 
-    public float latitude;
-    public float longitude;
+    public double latitude;
+    public double longitude;
 
 
 
-    public PhotoObject(int idname, String caption, String date, String source, float latitude, float longitude){
+    public PhotoObject(int idname, String caption, String date, String source, double latitude, double longitude){
         this.caption = caption;
         this.date = date;
         this.idname = idname;
